refactor: cell bottom border 추가 및 extension 파일분리 #6
