+++ OpenMarket/OpenMarket/Extension/Double+extension.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+extension Double {
+    func formattedPrice() -> String {
+        let numberFormatter = NumberFormatter()
+        numberFormatter.numberStyle = .decimal
+        
+        return numberFormatter.string(for: self) ?? ""
+    }
+}
