+++ OpenMarket/OpenMarket/Extension/String+extension.swift
@@ -1,0 +1,10 @@
+import UIKit
+
+extension String {
+    func strikeThrough() -> NSAttributedString {
+        let attributeString = NSMutableAttributedString(string: self)
+        attributeString.addAttribute(NSAttributedString.Key.strikethroughStyle, value: NSUnderlineStyle.single.rawValue, range: NSMakeRange(0, attributeString.length))
+        
+        return attributeString
+    }
+}
